feat(pagination): refactor structure of project
