+++ src/shared/components/pagination-ui-mode/constants/constants.ts
