+++ src/shared/components/pagination-ui-mode/exam/exam.module.css
