+++ src/shared/components/pagination-ui-mode/infinitive/infinitive.module.css
