+++ src/shared/components/pagination-ui-mode/random/random.module.css
