+++ src/shared/components/pagination-ui-mode/utils/swapQuestion.ts
